Move swagger to middleware directory
